+++ src/org/usfirst/frc/team1923/robot/commands/vision/VisionPegAlignCommand.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team1923.robot.commands.visionCommands;
+package org.usfirst.frc.team1923.robot.commands.vision;
 
 import org.usfirst.frc.team1923.robot.Robot;
 import org.usfirst.frc.team1923.robot.RobotMap;
@@ -8,21 +8,17 @@ import org.usfirst.frc.team1923.robot.RobotMap;
 import edu.wpi.first.wpilibj.command.Command;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
-/**
- *
- */
-public class TeleopVisionFeederAlignCommand extends Command {
+public class VisionPegAlignCommand extends Command {
 
     public double power, turn;
-    public boolean found;
+    // public boolean found;
     public boolean aligned;
     // private Timer time;
 
-    public TeleopVisionFeederAlignCommand() {
+    public VisionPegAlignCommand() {
         // Use requires() here to declare subsystem dependencies
         // eg. requires(chassis);
         requires(Robot.visionSubSys);
-        requires(Robot.driveSubSys);
     }
 
     // Called just before this Command runs the first time
@@ -39,37 +35,39 @@ public class TeleopVisionFeederAlignCommand extends Command {
         // new VisionAlignCommand(); TODO: Change to only run when needed to not
         // waste processor cycles
         Robot.visionSubSys.refresh();
-        if (Robot.visionSubSys.dist >= RobotMap.FEEDER_DIST) {
-            if (Robot.visionSubSys.turn < -1) {
-                power = 0;
-                turn = 0;
-                Robot.visionSubSys.found = false;
-            } else {
-                if (Robot.visionSubSys.dist >= 30)
-                    power = 0.4;
-                else power = 0.2;
-                // power=0;
-                Robot.visionSubSys.found = true;
-                turn = Robot.visionSubSys.turn;
-            }
-
-            Robot.driveSubSys.auto(power, turn);
-
+        if (Robot.visionSubSys.turn < -1) {
+            power = 0;
+            turn = 0;
+            Robot.visionSubSys.found = false;
         } else {
-            // Put SmartDashboard indicator
-            if (found && Robot.visionSubSys.dist <= RobotMap.FEEDER_DIST)
-                aligned = true;
-            else aligned = false;
-
-            SmartDashboard.putBoolean("Found: ", found);
-            SmartDashboard.putBoolean("Aligned and Ready to Drop: ", aligned);
+            if (Robot.visionSubSys.dist >= 30)
+                power = 0.45;
+            else power = 0.2;
+            // power=0;
+            Robot.visionSubSys.found = true;
+            turn = Robot.visionSubSys.turn;
         }
+
+        SmartDashboard.putNumber("Power", power);
+
+        // Testing
+        // System.out.println("Power: " + power + " Turn: " + turn);
+
+        Robot.driveSubSys.auto(power, turn);
+        if (Robot.visionSubSys.dist <= RobotMap.PEG_DIST)
+            aligned = true;
+        else aligned = false;
+
+        SmartDashboard.putBoolean("Found: ", Robot.visionSubSys.found);
+        SmartDashboard.putBoolean("Aligned and Ready to Drop: ", aligned);
     }
 
     // Make this return true when this Command no longer needs to run execute()
     @Override
     protected boolean isFinished() {
-        return Robot.visionSubSys.dist <= RobotMap.FEEDER_DIST;
+        if (Robot.visionSubSys.dist < RobotMap.PEG_DIST)
+            return true;
+        else return false;
     }
 
     // Called once after isFinished returns true
@@ -82,6 +80,7 @@ public class TeleopVisionFeederAlignCommand extends Command {
     // subsystems is scheduled to run
     @Override
     protected void interrupted() {
-
+        System.out.println("Interrupted!!!!!!!");
+        end();
     }
 }
